+++ src/components/Home.jsx
@@ -14,26 +14,53 @@ const Home = () => {
   const [podcastListFiltered, setPodcastListFiltered] = useState("");
   const [error, setError] = useState(null);
 
-  console.log("error", error, "filtered", filtered, "filtertext", filterText);
-
   let storagedList = localStorage.getItem("podcastList");
 
   useEffect(() => {
     (async () => {
       if (storagedList) {
         setPodcastArray(JSON.parse(storagedList));
+        setPodcastListFiltered(JSON.parse(storagedList));
         setLoaded(true);
       } else {
         try {
           await new Promise((resolve) => setTimeout(resolve, 500));
           setPodcastArray(podcastList);
+          setPodcastListFiltered(podcastList);
           setLoaded(true);
         } catch (error) {
           console.log("Error bringing data to the component", error);
         }
       }
     })();
+  }, []);
+
+  useEffect(() => {
+    const podcastFilteredArray = podcastList.filter((element) => {
+      let elementName = element.name;
+      let authorName = element.artist;
+
+      if (filterText) {
+        if (
+          elementName.toLowerCase().includes(filterText.toLowerCase()) ||
+          authorName.toLowerCase().includes(filterText.toLowerCase())
+        ) {
+          return element;
+        }
+      }
+    });
+    setPodcastListFiltered(podcastFilteredArray);
   }, [filterText]);
+
+  useEffect(() => {
+    if (filtered) {
+      if (podcastListFiltered.length === 0) {
+        setError(true);
+      } else if (podcastListFiltered.length !== 0) {
+        setError(false);
+      }
+    }
+  }, [podcastListFiltered]);
 
   return (
     <>
@@ -42,21 +69,14 @@ const Home = () => {
         <InputFilter setFilterText={setFilterText} setFiltered={setFiltered} />
         <ListCounter
           error={error}
-          podcastListFiltered={podcastListFiltered}
-          podcastArray={podcastArray}
+          podcastListFilteredLength={podcastListFiltered.length}
+          podcastArrayLength={podcastArray.length}
           filtered={filtered}
         />
       </header>
       <div>
         {filtered ? (
-          <PodcastCardsFiltered
-            filterText={filterText}
-            podcastListFiltered={podcastListFiltered}
-            setPodcastListFiltered={setPodcastListFiltered}
-            error={error}
-            setError={setError}
-            filtered={filtered}
-          />
+          <PodcastCardsFiltered podcastListFiltered={podcastListFiltered} error={error} />
         ) : (
           <PodcastCards podcastArray={podcastArray} loaded={loaded} />
         )}
